+++ Modules/ADUtils/UIView+NibLoader.swift
@@ -9,6 +9,8 @@
 import Foundation
 import UIKit
 
+// swiftlint:disable force_unwrapping
+
 /// NibLoadable is a semantic protocol defining an object instanciatable from a nib
 public protocol NibLoadable {}
 
@@ -19,7 +21,8 @@ extension NibLoadable {
 
     /**
      Instanciate a Self object from a nib
-     - parameter nibNameOrNil: A custom nib name, if nil Self name is used as nibs are generally demangled, default is nil
+     - parameter nibNameOrNil: A custom nib name, if nil Self name is used as nibs are generally demangled,
+     default is nil
      - parameter owner: The nib owner, default is nil
      - parameter bundle: The nib bundle, default is main bundle
      - note: Crashes if unable to instanciate the objet, meaning if there is no matching nib
@@ -34,8 +37,8 @@ extension NibLoadable {
     // MARK: - Private
 
     private static func ad_fromNib(_ nibNameOrNil: String? = nil,
-                                  owner: AnyObject? = nil,
-                                  bundle: Bundle = Bundle.main) -> Self? {
+                                   owner: AnyObject? = nil,
+                                   bundle: Bundle = Bundle.main) -> Self? {
         var view: Self?
         let name: String
         if let nibName = nibNameOrNil {