+++ Modules/ADUtils_objc/String+AttributedFormat_objc.swift
@@ -8,13 +8,16 @@
 
 import Foundation
 
-//???: (Benjamin Lavialle) 12/10/16 extensions over String are not usable in objc, this is why we need an extension on NSString
+// swiftlint:disable line_length
+
+// (Benjamin Lavialle) 12/10/16 extensions over String are not usable in objc,
+// this is why we need an extension on NSString
 
 extension NSString {
 
     @objc public func adobjc_attributedString(arguments: [String],
-                                        defaultAttributes: [NSAttributedString.Key: Any],
-                                        formatAttributes: [NSAttributedString.Key: Any]) -> NSAttributedString? {
+                                              defaultAttributes: [NSAttributedString.Key: Any],
+                                              formatAttributes: [NSAttributedString.Key: Any]) -> NSAttributedString? {
         return (self as String).attributedString(
             arguments: arguments,
             defaultAttributes: defaultAttributes,
@@ -23,8 +26,8 @@ extension NSString {
     }
 
     @objc public func adobjc_attributedString(arguments: [String],
-                                        defaultAttributes: [NSAttributedString.Key: Any],
-                                        differentFormatAttributes: [[NSAttributedString.Key: Any]]) -> NSAttributedString? {
+                                              defaultAttributes: [NSAttributedString.Key: Any],
+                                              differentFormatAttributes: [[NSAttributedString.Key: Any]]) -> NSAttributedString? {
         return (self as String).attributedString(
             arguments: arguments,
             defaultAttributes: defaultAttributes,